+++ packages/@straw-hat/fetcher/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  transform: {
-    '^.+\\.(t|j)sx?$': ['@swc/jest'],
-  },
-  testEnvironment: 'jsdom',
-  setupFilesAfterEnv: ['<rootDir>/tests/jest/jest.setup.ts'],
-};