+++ packages/@straw-hat/fetcher/src/helpers.ts
@@ -6,6 +6,12 @@ function isBlob(value: any): value is Blob {
   return value instanceof Blob;
 }
 
+// TODO: Support other types of body:
+// 1. json
+// 2. text
+// 3. blob
+// 4. arrayBuffer
+// 5. formData
 export function getRequestBody(body: any): BodyInit | undefined {
   if (body === undefined) {
     return undefined;