+++ packages/@straw-hat/fetcher/tests/vitest/openapi/index.test.ts
@@ -1,10 +1,11 @@
+import { expect, describe, test } from 'vitest';
 import { createUrlPath } from '../../../src/openapi';
 
 type PathParam = { article_id: number; comment_id: number };
 type QueryParam = { sort: 'asc' | 'desc'; sort_by: string };
 
 describe('createUrlPath', () => {
-  it('replaces the path params', () => {
+  test('replaces the path params', () => {
     const value = createUrlPath<PathParam, never>('/articles/{article_id}/comments/{comment_id}', {
       path: {
         article_id: 50,
@@ -15,7 +16,7 @@ describe('createUrlPath', () => {
     expect(value).toBe(expected);
   });
 
-  it('adds the query params to the url', () => {
+  test('adds the query params to the url', () => {
     const value = createUrlPath<PathParam, QueryParam>('/articles/{article_id}/comments/{comment_id}', {
       path: { article_id: 50, comment_id: 10 },
       query: { sort: 'asc', sort_by: 'title' },