+++ packages/@straw-hat/react-query-fetcher/src/use-fetcher-mutation.ts
@@ -1,18 +1,22 @@
-import type { Fetcher } from '@straw-hat/fetcher/dist';
-import type { UseMutationOptions } from '@tanstack/react-query';
-import { useMutation } from '@tanstack/react-query';
-import type { Endpoint, WithOptions } from './types';
+import { Fetcher } from '@straw-hat/fetcher';
+import { UseMutationOptions, useMutation } from '@tanstack/react-query';
+import { Endpoint, WithOptions } from './types.js';
+
+type Options<TData, TError, TVariables> = Omit<UseMutationOptions<TData, TError, TVariables>, 'mutationFn'>;
 
 export type UseFetcherMutationArgs<TData, TError, TVariables> = {
   endpoint: Endpoint<TData, TVariables>;
-  options?: UseMutationOptions<TData, TError, TVariables>;
+  options?: Options<TData, TError, TVariables>;
 };
 
 export function useFetcherMutation<TData = unknown, TError = unknown, TVariables = void>(
   client: Fetcher,
   args: UseFetcherMutationArgs<TData, TError, TVariables>
 ) {
-  return useMutation<TData, TError, WithOptions<TVariables>>(function (params) {
-    return args.endpoint(client, params);
-  }, args.options);
+  return useMutation<TData, TError, WithOptions<TVariables>>({
+    ...args.options,
+    mutationFn(params) {
+      return args.endpoint(client, params);
+    }
+  });
 }