+++ packages/@straw-hat/react-query-fetcher/src/use-fetcher-query.ts
@@ -1,13 +1,14 @@
-import type { Fetcher } from '@straw-hat/fetcher/dist';
-import type { UseQueryOptions } from '@tanstack/react-query';
-import { useQuery } from '@tanstack/react-query';
-import type { Endpoint, QueryKey } from './types';
+import {Fetcher} from '@straw-hat/fetcher';
+import {UseQueryOptions, useQuery, QueryKey} from '@tanstack/react-query';
+import {Endpoint} from './types.js';
+
+type Options<TQueryFnData, TError, TData> = Omit<UseQueryOptions<TQueryFnData, TError, TData>, 'queryKey' | 'queryFn'>
 
 export type UseFetcherQueryArgs<TQueryFnData, TError, TData, TParams> = {
   queryKey: QueryKey;
   endpoint: Endpoint<TQueryFnData, TParams>;
   params?: TParams;
-  options?: UseQueryOptions<TQueryFnData, TError, TData>;
+  options?: Options<TQueryFnData, TError, TData>;
 };
 
 export function createQueryKey<TParams = unknown>(queryKey: QueryKey, params?: TParams): QueryKey {
@@ -24,18 +25,18 @@ export function useFetcherQuery<TQueryFnData = unknown, TError = unknown, TData 
   const queryKey = createQueryKey(args.queryKey, args.params);
   const params = args.params ?? {};
 
-  return useQuery<TQueryFnData, TError, TData>(
+  return useQuery<TQueryFnData, TError, TData>({
+    ...args.options,
     queryKey,
-    function queryFn(context: { signal?: AbortSignal }) {
+    queryFn(context: { signal?: AbortSignal }) {
       return args.endpoint(
         client,
         // @ts-ignore ¯\_(ツ)_/¯
         {
           ...params,
-          options: { signal: context.signal },
+          options: {signal: context.signal},
         }
       );
     },
-    args.options
-  );
+  });
 }